split utils into new metrics module
